Remove target="_blank" from external links
It's bad for accessibility: https://www.a11yproject.com/checklist/#identify-links-that-open-in-a-new-tab-or-window

The markdown render doesn't add it by default anyway. Later we might add
rel="noreferrer" to all external links (which implies noopener), but
that should be done as a HTML post-processing for the full page to cover
both the layout and markdown.